agent/STUDYFORGE/requirements_gemini.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai>=0.7.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic>=2.0.0
